@@ -76,26 +76,26 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@eslint/js": "^9.23.0",
-    "@vitejs/plugin-react": "^4.3.4",
-    "coveralls-next": "^4.2.2",
-    "eslint": "^9.24.0",
-    "eslint-plugin-mocha": "^10.5.0",
+    "@eslint/js": "^9.39.1",
+    "@vitejs/plugin-react": "^4.7.0",
+    "coveralls-next": "^6.0.0",
+    "eslint": "^9.39.1",
+    "eslint-plugin-mocha": "^11.2.0",
     "eslint-plugin-react": "^7.37.5",
-    "globals": "^16.0.0",
-    "happy-dom": "^20.0.2",
-    "mocha": "^11.1.0",
+    "globals": "^16.5.0",
+    "happy-dom": "^20.0.10",
+    "mocha": "^11.7.5",
     "mocha-lcov-reporter": "^1.3.0",
     "node-deb": "^0.10.7",
     "node-env-run": "^4.0.2",
-    "nodemon": "^3.1.9",
+    "nodemon": "^3.1.11",
     "npm-run-all": "^4.1.5",
     "nyc": "^17.1.0",
     "pino-colada": "^2.2.2",
-    "pino-http": "^10.3.0",
+    "pino-http": "^11.0.0",
     "sinon": "^21.0.0",
     "vite-plugin-eslint": "^1.8.1",
-    "vitest": "^3.1.1"
+    "vitest": "^4.0.10"
   },
   "nyc": {
     "all": true,
